feat: интеграция с реальным libbox.aar 1.12.9 (native-сборка)
- HydraPlatformInterface переписан под фактический API libbox 1.12.9
  (javap-сверка с собранным .aar): findConnectionOwner -> uid, монитор
  интерфейса через InterfaceUpdateListener + ConnectivityManager,
  getInterfaces/systemCertificates итераторы, packageNameByUid/uidByPackageName,
  sendNotification, clearDNSCache, autoDetectInterfaceControl
- SingBoxConfigBuilder: DNS-схема 1.12 (type: https/local), tun без
  auto_route/strict_route (маршруты владеет VpnService.Builder)
- build.gradle.kts: libXray.aar опционален (условная зависимость),
  комментарии восстановлены после mojibake
- SstpCore: фикс типа таймаута connect()
- docs/BUILD.md: сверено — сборка libbox требует Go 1.25 + -checklinkname=0
  (workaround sing-box#3233 / golang#70508)

Проверено: :app:assembleNativeDebug — BUILD SUCCESSFUL (libbox 1.12.9,
теги with_gvisor,with_quic,with_wireguard,with_utls,with_clash_api,with_conntrack)

+++ app/src/native/java/ru/gidravpn/hydra/vpn/core/SstpCore.kt
@@ -90,7 +90,7 @@ class SstpCore : VpnCore {
         onLog("SSTP: подключение к ${profile.address}:${profile.port}…")
         val raw = Socket()
         SocketGuard.protect(raw) // иначе TLS-трафик уйдёт в собственный tun
-        raw.connect(InetSocketAddress(profile.address, profile.port), CONNECT_TIMEOUT_MS)
+        raw.connect(InetSocketAddress(profile.address, profile.port), CONNECT_TIMEOUT_MS.toInt())
         raw.tcpNoDelay = true
 
         val ssl = createTlsSocket(raw, profile, allowInsecure, extra.optString("sni"))
